first pass at upgrading contracts to 0.8

+++ pkg/solidity-utils/contracts/helpers/VaultHelpers.sol
@@ -23,6 +23,6 @@ library VaultHelpers {
     function toPoolAddress(bytes32 poolId) internal pure returns (address) {
         // 12 byte logical shift left to remove the nonce and specialization setting. We don't need to mask,
         // since the logical shift already sets the upper bits to zero.
-        return address(uint256(poolId) >> (12 * 8));
+        return address(uint160(uint256(poolId) >> (12 * 8)));
     }
 }
